requirements/base.txt
# Configuration
django-environ==0.4.0
django-secure==1.0.1
whitenoise==2.0.4


# Forms
django-braces==1.8.1
django-crispy-forms==1.5.2
django-floppyforms==1.5.2

# Models and persistent storage
django-model-utils==2.3.1
mysqlclient==1.3.6

# Images
Pillow==3.0.0

# For user registration, either via email or social
# Well-built with regular release cycles!
django-allauth==0.23.0

# Unicode slugification
unicode-slugify==0.1.3
django-autoslug==1.9.3

# Time zones support
pytz==2015.6

# Views
django-tables2==1.0.4
django-filter==0.11.0
django-autocomplete-light==2.2.10
django-extra-views==0.7.1

# Your custom requirements go here
git+https://github.com/ad-m/django-atom.git@master#egg=django-atom
git+https://github.com/ad-m/django-teryt-tree.git@master#egg=django-teryt-tree
django-states==1.6.5
django-bootstrap-pagination==1.5.1
cached-property==1.2.0
django_localflavor==1.1
jsonfield==1.0.3
django-bleach==0.3.0
djmail==0.11.0
